refactor breath list screen
