+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/lists/breath/BreathContainer.kt
@@ -1,0 +1,64 @@
+package com.example.zeldacompendium.presentation.ui.lists.breath
+
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.focus.FocusManager
+import androidx.compose.ui.platform.LocalFocusManager
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavController
+import com.example.zeldacompendium.domain.CompendiumFilter
+import com.example.zeldacompendium.domain.CompendiumFilterImpl
+import com.example.zeldacompendium.presentation.ui.components.SetBackgroundImage
+import com.example.zeldacompendium.presentation.ui.components.SetFrame
+import com.example.zeldacompendium.presentation.ui.lists.breath.sections.BottomBarSection
+import com.example.zeldacompendium.presentation.ui.lists.breath.sections.BreathListSection
+import com.example.zeldacompendium.presentation.ui.lists.breath.sections.TopBarSection
+
+@Composable
+fun BreathContainer(
+   gameId: Int?,
+   navController: NavController,
+   viewModel: BreathViewModel = hiltViewModel(),
+   focusManager: FocusManager = LocalFocusManager.current,
+   compendiumFilter: CompendiumFilter = CompendiumFilterImpl()
+) {
+   var selectedIndex by remember { mutableIntStateOf(0) }
+   val isLoading by remember { viewModel.isLoading }
+   val isError by remember { viewModel.loadError }
+   var searchText by remember { mutableStateOf("") }
+   val filteredList = compendiumFilter.filterCompendiumList(viewModel.compendiumList.value, selectedIndex)
+
+   Scaffold(
+      topBar = {
+         TopBarSection(navController = navController)
+      },
+      bottomBar = {
+         BottomBarSection(
+            isError = isError,
+            onSearch = { query ->
+               searchText = query
+               viewModel.searchCompendium(query)
+            },
+            onCategorySelected = { newSelectedIndex ->
+               focusManager.clearFocus()
+               selectedIndex = newSelectedIndex
+            }
+         )
+      }
+   ) { padding ->
+      SetBackgroundImage()
+      SetFrame()
+      BreathListSection(
+         padding = padding,
+         gameId = gameId,
+         filteredList = filteredList,
+         isLoading = isLoading,
+         isError = isError
+      )
+   }
+}
