+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/lists/breath/sections/BreathListSection.kt
@@ -1,8 +1,10 @@
-package com.example.zeldacompendium.presentation.ui.lists.breath.components
+package com.example.zeldacompendium.presentation.ui.lists.breath.sections
 
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
@@ -16,14 +18,41 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.zeldacompendium.data.models.CompendiumListEntry
 import com.example.zeldacompendium.presentation.ui.components.RetrySection
-import com.example.zeldacompendium.presentation.ui.lists.ImageList
-import com.example.zeldacompendium.presentation.ui.lists.breath.CompendiumBreathViewModel
+import com.example.zeldacompendium.presentation.ui.lists.ListHeaderImageSection
+import com.example.zeldacompendium.presentation.ui.lists.breath.BreathViewModel
+import com.example.zeldacompendium.presentation.ui.lists.breath.sections.items.CompendiumItemBreathEmpty
+import com.example.zeldacompendium.presentation.ui.lists.breath.sections.items.CompendiumItemBreath
+
+@Composable
+fun BreathListSection(
+   padding: PaddingValues,
+   gameId: Int?,
+   filteredList: List<CompendiumListEntry>,
+   isLoading: Boolean,
+   isError: String
+) {
+   Box(
+      modifier = Modifier.padding(padding)
+   ) {
+      if (filteredList.isEmpty() && !isLoading && isError.isEmpty()) {
+         Column(modifier = Modifier.padding(16.dp)) {
+            ListHeaderImageSection(gameId = gameId)
+            CompendiumItemBreathEmpty()
+         }
+      } else {
+         CompendiumList(
+            gameId = gameId,
+            compendiumList = filteredList
+         )
+      }
+   }
+}
 
 @Composable
 fun CompendiumList(
    gameId: Int?,
    compendiumList: List<CompendiumListEntry>,
-   viewModel: CompendiumBreathViewModel = hiltViewModel(),
+   viewModel: BreathViewModel = hiltViewModel(),
 ) {
 
    val loadError by remember { viewModel.loadError }
@@ -32,7 +61,7 @@ fun CompendiumList(
    LazyColumn(contentPadding = PaddingValues(16.dp)) {
       val itemCount = compendiumList.size
       item {
-         ImageList(gameId = gameId)
+         ListHeaderImageSection(gameId = gameId)
       }
       items(itemCount) {
          if (it >= itemCount && !isLoading) {
